+++ cuketmon/src/App.css
@@ -6,90 +6,8 @@
 }
 
 html, body, *{
-  font-family: 'neodgm', sans-serif;
-}
-/*로그인화면 CSS*/
-html {
-  background-color: black;
-  font-family: "neodgm", sans-serif;
-}
-
-.App {
+  margin: 0, auto;
+  font-family: 'neodgm';
   width: 100vw;
   height: 100vh;
-  padding-bottom: 10%;
-  overflow: hidden;
-  background-image: url("/public/LoginPage/background.png");
-  background-size: cover;
-  background-repeat: no-repeat;
-}
-
-#webLogo {
-  padding-left: 23%;
-  width: 50%;
-  margin-top: 10%;
-}
-
-.buttonContainer .buttonImage {
-  position: absolute;
-  top: 67%;
-  left: 41%;
-  width: 22%;
-}
-
-.buttonContainer .buttonImage:hover {
-  content: url("/public/button-hover.png");
-  width: 22%;
-  z-index: 0;
-}
-
-.buttonText {
-  position: absolute;
-  top: 75.9%;
-  left: 49.5%;
-  font-size: 15px;
-  font-weight: bold;
-  color: white;
-  z-index: 1;
-}
-
-/*모바일 반응형*/
-@media (max-width: 767px) {
-  #root,
-  body {
-    height: 100%;
-    padding: 0;
-  }
-
-  .App {
-    height: 100%;
-    padding: 0;
-  }
-
-  #webLogo {
-    width: 100%;
-    padding-top: 40%;
-    padding-left: 0;
-  }
-
-  .buttonContainer .buttonImage {
-    position: absolute;
-    width: 53%;
-    left: 28%;
-  }
-
-  .buttonContainer .buttonImage:hover {
-    content: url("/public/button-hover.png");
-    width: 53%;
-  }
-
-  .buttonText {
-    position: absolute;
-    top: 72.2%;
-    left: 48%;
-    font-size: 9px;
-    font-weight: bold;
-    color: white;
-    z-index: 1;
-  }
 }
